refactor(rules): extract shared parser/newline helpers into support module
Removes two near-identical chunks of code:

  * The `ProtectedLineCollector` + `protected_scalar_lines` pair was
    duplicated between empty_lines.rs and trailing_spaces.rs (one
    protected every scalar, the other only literal/folded/multi-line
    double-quoted). They now share a single generic helper in
    support/line_syntax.rs that takes a `Fn(ScalarStyle, Span) -> bool`
    filter. The block-scalar end-exclusive adjustment lives in one
    place.
  * The byte-for-byte identical `first_newline` helper in
    document_start.rs and document_end.rs is now a shared
    `buffer_newline` in support/line_syntax.rs.

Net source delta: -24 lines (one extraction is bigger than its two
sites, the other is straight deduplication). Coverage stays at 100%
and all 1012 tests pass without changes; this is purely a structural
refactor.

Refs #221
