+++ src/rules/trailing_spaces.rs
@@ -11,11 +11,11 @@
 //! parse time, so those lines remain fixable. The protected line set is
 //! computed via `saphyr_parser`, so the rule bails (returns `None`) when
 //! the buffer cannot be parsed.
-use std::collections::HashSet;
+use saphyr_parser::ScalarStyle;
 
-use saphyr_parser::{Event, Parser, ScalarStyle, Span, SpannedEventReceiver};
-
-use crate::rules::support::line_syntax::split_lines_preserve_endings;
+use crate::rules::support::line_syntax::{
+    protected_scalar_lines, split_lines_preserve_endings,
+};
 
 pub const ID: &str = "trailing-spaces";
 pub const MESSAGE: &str = "trailing spaces";
@@ -85,7 +85,11 @@ fn process_line(
 
 #[must_use]
 pub fn fix(buffer: &str) -> Option<String> {
-    let protected = protected_scalar_lines(buffer)?;
+    let protected = protected_scalar_lines(buffer, |style, span| match style {
+        ScalarStyle::Literal | ScalarStyle::Folded => true,
+        ScalarStyle::DoubleQuoted => span.end.line() > span.start.line(),
+        _ => false,
+    })?;
     let mut output = String::with_capacity(buffer.len());
     let mut changed = false;
 
@@ -105,41 +109,4 @@ pub fn fix(buffer: &str) -> Option<String> {
     }
 
     changed.then_some(output)
-}
-
-fn protected_scalar_lines(buffer: &str) -> Option<HashSet<usize>> {
-    let mut parser = Parser::new_from_str(buffer);
-    let mut collector = ProtectedLineCollector {
-        protected: HashSet::new(),
-    };
-    parser.load(&mut collector, true).ok()?;
-    Some(collector.protected)
-}
-
-struct ProtectedLineCollector {
-    protected: HashSet<usize>,
-}
-
-impl SpannedEventReceiver<'_> for ProtectedLineCollector {
-    fn on_event(&mut self, event: Event<'_>, span: Span) {
-        if let Event::Scalar(_, style, _, _) = event {
-            let start = span.start.line();
-            let end = span.end.line();
-            let needs_protect = match style {
-                ScalarStyle::Literal | ScalarStyle::Folded => true,
-                ScalarStyle::DoubleQuoted => end > start,
-                _ => false,
-            };
-            if needs_protect {
-                let last = if span.end.col() == 0 && end > start {
-                    end - 1
-                } else {
-                    end
-                };
-                for line in start..=last {
-                    self.protected.insert(line);
-                }
-            }
-        }
-    }
 }
